+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+2.0.5
+-----
+- Added customer hasCustomerAccount property
+- Added new shopware 5.2 attribute management
+- Added translated product attributes (pull)
+- Added Wawi missing product default price work around
+- Fixed product base price
+
 2.0.4
 -----
 - Changing customer order item gross price to 4 decimal places